correções gerais
