--- docker-compose.yaml
@@ -1,9 +1,0 @@
-version: "3.8"
-
-services:
-  database:
-    image: "postgres:16.0-alpine3.18"
-    env_file:
-      - ./.env
-    ports:
-      - "5432:5432"
