--- logs/flask_app.log.1
@@ -1,2 +1,0 @@
-[2024/08/08 04:16:03 AM +0000] INFO in __init__: Flask App Startup
-[2024/08/08 04:16:03 AM +0000] INFO in __init__: Flask App Startup
