--- logs/flask_app.log.3
@@ -1,1 +1,0 @@
-[2024/08/07 02:55:11 AM +0000] INFO in __init__: Flask App Startup
