+++ src/app/api/__init__.py
@@ -1,0 +1,3 @@
+from flask import Blueprint
+
+api = Blueprint('api', __name__)
