+++ src/tests/test_views.py
@@ -41,4 +41,3 @@ def test_account(logged_in_client):
     assert '<title>Saúde + Solidária • Conta</title>' in response.data \
         .decode('utf-8')
     assert b'<p>Bem vindo(a) Test User</p>' in response.data
-
